refactor(Refactorización-de-la-estructura-de-las-carpetas): cambiar orden entre las carpetas para una mejor compresión del código
